pyproject.toml
[project]
name = "automaton-auditor"
version = "0.1.0"
description = "LangGraph-based multi-agent repository auditing framework."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Automaton Auditor Contributors" }
]
dependencies = [
  "langgraph>=1.0.9,<2.0.0",
  "langchain>=1.2.10,<2.0.0",
  "langchain-ollama>=1.0.1,<2.0.0",
  "langchain-google-genai>=4.2.1,<5.0.0",
  "python-dotenv>=1.2.1,<2.0.0",
  "pydantic>=2.12.5,<3.0.0",
  "typing-extensions>=4.15.0,<5.0.0",
  "PyPDF2>=3.0.1,<4.0.0",
  "pymupdf>=1.24.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "ruff>=0.8.0,<1.0.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
  "pytest>=8.3.0,<9.0.0",
  "ruff>=0.8.0,<1.0.0",
]
